# Add support to detect Neoverse V2 cores